Generalize packet structure and add interfaces/abstract classes
