+++ src/main/java/edu/ntnu/idi/idatt/boardgame/App.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idi.idatt.boardgame;
 
-import edu.ntnu.idi.idatt.boardgame.controller.GameController;
+import edu.ntnu.idi.idatt.boardgame.controller.SnakesAndLaddersController; // Updated import
 import edu.ntnu.idi.idatt.boardgame.view.GameView;
 import javafx.application.Application;
 import javafx.scene.Scene;
@@ -9,15 +9,12 @@ import javafx.stage.Stage;
 public class App extends Application {
   @Override
   public void start(Stage stage) {
-    GameController gameController = new GameController();
-
+    SnakesAndLaddersController gameController = new SnakesAndLaddersController(2); // Updated
     GameView gameView = new GameView(gameController, gameController.getGameBoard());
-
-    gameController.setGameView(gameView);
 
     Scene scene = new Scene(gameView.getRoot(), 700, 800);
     stage.setScene(scene);
-    stage.setTitle("JavaFX Maven App");
+    stage.setTitle("Snakes and Ladders"); // Updated title for clarity
     stage.show();
   }
