+++ src/main/java/edu/ntnu/idi/idatt/boardgame/controller/GameObserver.java
@@ -1,0 +1,6 @@
+package edu.ntnu.idi.idatt.boardgame.controller;
+
+public interface GameObserver {
+    void update(String message);
+    void gameFinished(int winnerId);
+}
