+++ src/main/java/edu/ntnu/idi/idatt/boardgame/controller/SnakesAndLaddersController.java
@@ -1,0 +1,56 @@
+package edu.ntnu.idi.idatt.boardgame.controller;
+
+import edu.ntnu.idi.idatt.boardgame.action.CompositeAction;
+import edu.ntnu.idi.idatt.boardgame.action.RollAction;
+import edu.ntnu.idi.idatt.boardgame.domain.board.GameBoard;
+import edu.ntnu.idi.idatt.boardgame.domain.board.SnakesAndLaddersBoard;
+import edu.ntnu.idi.idatt.boardgame.domain.dice.Dice;
+import edu.ntnu.idi.idatt.boardgame.domain.player.Player;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.IntStream;
+import javafx.scene.paint.Color;
+
+public class SnakesAndLaddersController extends GameController {
+    private final int numberOfPlayers;
+    private final List<Color> playerColors = List.of(Color.RED, Color.BLUE, Color.GREEN, Color.YELLOW, Color.ORANGE, Color.PURPLE);
+
+    public SnakesAndLaddersController(int numberOfPlayers) {
+        super(new SnakesAndLaddersBoard(), new Dice(2)); // Pass only gameBoard and dice
+        this.numberOfPlayers = numberOfPlayers;
+        initialize(numberOfPlayers); // Call initialize after fields are set
+    }
+
+    @Override
+    protected Map<Integer, Player> createPlayers(int numberOfPlayers) {
+        Map<Integer, Player> players = new HashMap<>();
+        IntStream.rangeClosed(1, numberOfPlayers).forEach(playerId -> {
+            Player player = new Player(playerId, playerColors.get(playerId - 1));
+            players.put(playerId, player);
+        });
+        return players;
+    }
+
+    @Override
+    protected String performTurn() {
+        CompositeAction currentAction = new CompositeAction();
+        currentAction.addAction(new RollAction(gameBoard, currentPlayer, dice));
+        return currentAction.execute();
+    }
+
+    @Override
+    protected boolean isGameOver() {
+        return currentPlayer.getPosition() == gameBoard.getBoardSize();
+    }
+
+    @Override
+    protected void onGameFinish() {
+        notifyGameFinished(currentPlayer.getId());
+    }
+
+    @Override
+    protected Player getNextPlayer() {
+        return players.get((currentPlayer.getId() % numberOfPlayers) + 1);
+    }
+}
