+++ src/main/java/edu/ntnu/idi/idatt/boardgame/domain/dice/DiceInterface.java
@@ -1,0 +1,5 @@
+package edu.ntnu.idi.idatt.boardgame.domain.dice;
+
+public interface DiceInterface {
+    int roll();
+}
